isolate response definition
