+++ src/suite/mod.rs
@@ -1,4 +1,5 @@
 pub mod report;
+pub mod response;
 pub mod request;
 pub mod setup;
 pub mod suite;
