+++ src/suite/response/mod.rs
@@ -1,0 +1,4 @@
+pub mod response_definition;
+pub mod response_headers;
+
+pub use response_definition::ResponseDefinition;
