+++ src/suite/test.rs
@@ -11,7 +11,7 @@ use crate::{
     matcher::{match_result::MatchResult, status_matcher::StatusMatcher, MatchCmp},
 };
 
-use super::request::RequestDefinition;
+use super::{response::ResponseDefinition, request::RequestDefinition};
 
 #[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
 pub struct Test {
@@ -80,68 +80,6 @@ impl fmt::Display for FailureReport {
         writeln!(f, "{}", self.match_result)?;
 
         Ok(())
-    }
-}
-
-#[derive(Debug, Clone, Serialize, Deserialize, PartialEq)]
-pub struct ResponseDefinition {
-    pub status: Option<StatusMatcher>,
-    pub headers: Option<HashMap<String, String>>,
-    pub body: Option<serde_json::Value>,
-}
-
-impl ResponseDefinition {
-    pub async fn from_response(response: reqwest::Response) -> Self {
-        let status = Some(StatusMatcher::Exact(response.status().as_u16()));
-        let header_map = response.headers().clone();
-        let headers = header_map
-            .into_iter()
-            .filter_map(|(header, value)| match header {
-                Some(header) => Some((header, value)),
-                None => None,
-            })
-            .map(|(header, value)| {
-                (
-                    header.as_str().to_string(),
-                    value.to_str().unwrap().to_string(),
-                )
-            })
-            .collect::<HashMap<String, String>>();
-
-        let headers = match headers.len() {
-            0 => None,
-            _ => Some(headers),
-        };
-
-        let body = match response.json::<serde_json::Value>().await {
-            Ok(body) => Some(body),
-            Err(_) => None,
-        };
-
-        ResponseDefinition {
-            status,
-            headers,
-            body,
-        }
-    }
-
-    pub fn compare(&self, other: &ResponseDefinition) -> TestResult {
-        match self.status.match_cmp(&other.status) {
-            MatchResult::Matches => {}
-            other => return TestResult::fail("Status does not match.", &other),
-        }
-
-        match self.headers.match_cmp(&other.headers) {
-            MatchResult::Matches => {}
-            other => return TestResult::fail("Headers do not match.", &other),
-        }
-
-        match self.body.match_cmp(&other.body) {
-            MatchResult::Matches => {}
-            other => return TestResult::fail("Body does not match.", &other),
-        }
-
-        return TestResult::Passed;
     }
 }
